Refactor document_chunker and tests
