+++ pyproject.toml
@@ -27,6 +27,9 @@ dependencies = [
 packages = [{ include = "ingestion_code", from = "src" }]
 
 
+[tool.poetry.group.test.dependencies]
+pytest-mock = "^3.14.1"
+
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
